Radar & Control System Updates
The radar system was updated to include dynamic range and to properly
calculate target position for a 2D radar in a 3D world.  This also meant
updating the control system because it was polling to fast and executing
a button press twice instead of once. Hud updates were included to show
new dynamic radar range and to attempt to show Z-axis info but it's not
currently working.